refactor: extract shared host-normalisation helpers to hostnorm.go
Add untagged hostnorm.go with unbracketHost, isWildcardListenHost, and
navigableHost so browser.go and tray.go share one implementation.
